Completed assignment

+++ Binomial.java
@@ -13,7 +13,6 @@ public class Binomial {
 	// Computes the Binomial function, basic version.
 	public static int binomial1(int n, int k) 
 	{
-		
 		if (k > n) 
         	return 0;
     
@@ -25,9 +24,9 @@ public class Binomial {
 
 		
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) 
+	public static long binomial(int n, int k) 
 	{
-		int[][] memo = new int[n + 1][k + 1];
+		long[][] memo = new long[n + 1][k + 1];
 
 		for (int i = 0; i <= n; i++) 
         	for (int j = 0; j <= k; j++) 
@@ -38,7 +37,7 @@ public class Binomial {
 	
 	
 
-	private static int binomial(int n, int k, int[][] memo) 
+	private static long binomial(int n, int k, long[][] memo) 
 	{
 		if (memo[n][k] != -1) {
 			return memo[n][k];
